uart: Change to direct driver that handles RX and TX
The previous driver included multiple nrf-related files that
just handled queues and buffers in a overly complicated way.

Since FreeRTOS in the end has its own queue handling, and
suitable stream_buffers for asynchrnous byte streams, it is
better to use them directly instead

This driver will integrate both RX and TX into the FreeRTOS
stream buffer interface, to use a bare minimum of resources
for transmission and reception.

The UARTE would be better in the case where no RX is expected,
but for a generic driver UART has better buffering